Harden client after full audit: retry-on-401, status checks, token safety
Bugs fixed (each pinned by a regression test in tests/audit_probes.rs):
- delete_blog_target ignored HTTP status: 500 returned Ok(()), 401
  surfaced as JsonParse; now routed through handle_response
- no retry on 401 despite README claims: send_authorized now forces a
  token refresh and retries once when the refresh flow is configured
  (multipart bodies stream and are not retried)
- derive(Debug) leaked bearer/refresh tokens; AuthState now has a
  manual redacting Debug impl
- get_posts returned more items than `limit` when the server
  over-delivered; now truncates and short-circuits limit=0
- negative expires_in no longer wraps into a huge token lifetime

New:
- ApiClient::refresh_token() getter so the rotated refresh token can
  be persisted across restarts
- tests/live_api.rs: ignored-by-default live contract canary against
  public anonymous Boosty data; caught Post.price being fractional
  (0.98) -> all price-like fields are now f64
- methodology tooling: CLAUDE.md, scripts/ci.ps1, .claude git gate

Models:
- price fields i32/u64 -> f64 (live evidence)
- colliding unnameable types renamed (PostsExtra, CommentsExtra,
  ShowcaseExtra, SearchTagsExtra, PromoCount); all model types are now
  re-exported from model.rs
- CurrencyPrices rub/usd get serde(default) like eur

Deps:
- reqwest 0.12 -> 0.13 (.form() moved behind the new form feature)
- tokio slimmed to features=[sync] for the library; macros/rt moved
  to dev-dependencies; rust_decimal drops unused macros feature
- Cargo.lock committed for reproducible CI

Cleanup: dead ApiError::NotAvailable/JsonParse variants removed, stale
error docs fixed, blind upstream retry test replaced with a real one
verifying token rotation (r1->r2->r3), comments-pagination flag
semantics documented in code (verified live: terminal page in any
order is exactly isFirst && isLast).

+++ src/model/subscription.rs
@@ -27,10 +27,10 @@ pub struct Subscription {
     pub parent_id: Option<u64>,
     /// Display name of the subscription.
     pub name: String,
-    /// Standard price (in base currency units).
-    pub price: u64,
+    /// Standard price (in base currency units). Fractional values occur on the live API.
+    pub price: f64,
     /// Custom price, if applied.
-    pub custom_price: u64,
+    pub custom_price: f64,
     /// Billing period in months.
     pub period: u8,
     /// Start timestamp (Unix epoch).
@@ -69,8 +69,8 @@ pub struct SubscriptionLevelInfo {
     pub id: u64,
     /// Name of the level.
     pub name: String,
-    /// Base price in main currency.
-    pub price: u64,
+    /// Base price in main currency. Fractional values occur on the live API.
+    pub price: f64,
     /// Price per currency (e.g., USD, EUR).
     pub currency_prices: HashMap<String, f64>,
     /// Whether the level has limited availability.
